Extract file line metadata partial
